requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
bidict==0.22.1
biopython==1.81
blinker==1.6.2
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.1.0
cycler==0.11.0
dnspython==2.3.0
eventlet==0.33.3
Flask==2.3.2
Flask-Cors==3.0.10
Flask-SocketIO==5.3.4
fonttools==4.40.0
frozenlist==1.3.3
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.3
matplotlib==3.7.1
multidict==6.0.4
numpy==1.25.0
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
pyparsing==3.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.4.1
python-socketio==5.8.0
pytz==2023.3
seaborn==0.12.2
six==1.16.0
tzdata==2023.3
uvicorn==0.22.0
websockets==11.0.3
Werkzeug==2.3.6
yarl==1.9.2
zipp==3.15.0
